requirements.txt
streamlit
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-auth
google-cloud-storage
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
moviepy==1.0.3
pydub
whisper
openai
python-dotenv
streamlit_webrtc
srt
reportlab
PyVimeo
pysrt
yt-dlp